@@ -10,22 +10,24 @@
     "test": "echo \"No test specified\" && exit 0"
   },
   "dependencies": {
-    "@quasar/extras": "^1.0.0",
-    "axios": "^0.18.1",
-    "core-js": "^3.6.5",
-    "csv-generate": "^3.3.0",
-    "csv-parse": "^4.15.0",
-    "ethers": "^5.0.26",
-    "quasar": "^1.0.0"
+    "@quasar/cli": "^1.2.2",
+    "@quasar/extras": "^1.12.4",
+    "axios": "^0.25.0",
+    "csv": "^6.0.5",
+    "ethers": "^5.5.4",
+    "quasar": "^2.5.2",
+    "stream": "^0.0.2",
+    "vue": "^3.0.0",
+    "vue-router": "^4.0.0"
   },
   "devDependencies": {
-    "@quasar/app": "^2.0.0",
-    "babel-eslint": "^10.0.1",
+    "@babel/eslint-parser": "^7.13.14",
+    "@quasar/app": "^3.3.2",
     "dotenv": "^8.2.0",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.9.0",
-    "eslint-loader": "^3.0.3",
-    "eslint-plugin-vue": "^6.1.2"
+    "eslint": "^7.14.0",
+    "eslint-config-prettier": "^8.1.0",
+    "eslint-plugin-vue": "^7.0.0",
+    "eslint-webpack-plugin": "^2.4.0"
   },
   "browserslist": [
     "last 10 Chrome versions",
@@ -39,7 +41,7 @@
     "last 5 Opera versions"
   ],
   "engines": {
-    "node": ">= 10.18.1",
+    "node": ">= 12.22.1",
     "npm": ">= 6.13.4",
     "yarn": ">= 1.21.1"
   }
