@@ -2,11 +2,11 @@
   <q-page class="" id="pageChainTransactions">
     <table-transactions
       :title="'Chain Transactions - ' + $store.taxYear"
-      :data="filtered"
+      :rows="filtered"
       :columns="columns"
     >
       <template v-slot:top-right>
-        <div style="min-width: 250px; display: inline-block;" class="q-mr-sm">
+        <div style="min-width: 250px; display: inline-block" class="q-mr-sm">
           <q-select
             filled
             v-model="$store.selectedAccounts"
@@ -31,7 +31,6 @@
 import { store } from "../boot/store";
 import { actions } from "../boot/actions";
 import { getChainTransactions, columns } from "../services/chain-tx-provider";
-import Vue from "Vue";
 import TableTransactions from "src/components/TableTransactions.vue";
 import { filterByAccounts, filterByYear } from "src/services/filter-service";
 
@@ -45,7 +44,7 @@ export default {
       page: 1,
       chainTransactions: Object.freeze([]),
       $store: store,
-      $actions: actions
+      $actions: actions,
     };
   },
   computed: {
@@ -54,7 +53,7 @@ export default {
 
       if (this.onlyShowUnNamed) {
         txs = txs.filter(
-          tx =>
+          (tx) =>
             tx.toName.substring(0, 2) == "0x" ||
             tx.fromName.substring(0, 2) == "0x"
         );
@@ -62,23 +61,23 @@ export default {
         txs = filterByAccounts(txs, this.$store.selectedAccounts, true);
       }
       return Object.freeze(txs);
-    }
+    },
   },
   methods: {
     async load() {
       const chainTransactions = await getChainTransactions();
-      Vue.set(this, "chainTransactions", Object.freeze(chainTransactions));
-    }
+      this.chainTransactions = Object.freeze(chainTransactions);
+    },
   },
   async created() {
     await this.load();
     store.onload = this.load;
   },
-  destroyed() {
+  unmounted() {
     store.onload = null;
   },
   mounted() {
     window.__vue_mounted = "pageChainTransactions";
-  }
+  },
 };
 </script>
