@@ -2,43 +2,44 @@
   <q-page class="" id="pageRunningBalances" ref="transactionPage">
     <table-transactions
       :title="'Running Balances - ' + $store.taxYear"
-      :data="filtered"
+      :rows="filtered"
       :columns="columns"
     >
       <template v-slot:top-right>
         <filter-account-asset></filter-account-asset>
-        <q-btn-dropdown stretch flat :label="balanceGrouping">
-          <q-list>
-            <q-item
-              v-for="n in groups"
-              :key="`x.${n}`"
-              clickable
-              v-close-popup
-              tabindex="0"
-              @click="balanceGrouping = n"
-            >
-              <q-item-label>{{ n }}</q-item-label>
-            </q-item>
-          </q-list>
-        </q-btn-dropdown>
+        <div style="display: inline-block" class="q-mr-sm">
+          <q-btn-dropdown stretch flat :label="balanceGrouping">
+            <q-list>
+              <q-item
+                v-for="n in groups"
+                :key="`x.${n}`"
+                clickable
+                v-close-popup
+                tabindex="0"
+                @click="balanceGrouping = n"
+              >
+                <q-item-label>{{ n }}</q-item-label>
+              </q-item>
+            </q-list>
+          </q-btn-dropdown>
+        </div>
       </template>
     </table-transactions>
   </q-page>
 </template>
 
 <script>
-import { store } from "../boot/store";
+//import { store } from "../boot/store";
 import { actions } from "../boot/actions";
 import {
   getRunningBalances,
-  columns
+  columns,
 } from "../services/running-balances-provider";
 import {
   filterByAccounts,
   filterByAssets,
-  filterByYear
+  filterByYear,
 } from "../services/filter-service";
-import Vue from "Vue";
 import FilterAccountAsset from "src/components/FilterAccountAsset.vue";
 import TableTransactions from "src/components/TableTransactions.vue";
 export default {
@@ -49,13 +50,13 @@ export default {
       balanceGrouping: "Detailed",
       runningBalances: Object.freeze([]),
       columns: columns,
-      $store: store,
-      $actions: actions
+      $store: this.$store,
+      $actions: actions,
     };
   },
   components: {
     FilterAccountAsset,
-    TableTransactions
+    TableTransactions,
   },
   computed: {
     filtered() {
@@ -68,8 +69,8 @@ export default {
         if (taxYear == "All") {
           taxYear = this.$store.taxYears[this.$store.taxYears.length - 2];
         }
-        txs = txs.filter(tx => {
-          return tx.accountEndingYears.findIndex(ey => ey == taxYear) > -1;
+        txs = txs.filter((tx) => {
+          return tx.accountEndingYears.findIndex((ey) => ey == taxYear) > -1;
         });
       }
       if (this.balanceGrouping == "Asset") {
@@ -77,8 +78,8 @@ export default {
         if (taxYear == "All") {
           taxYear = this.$store.taxYears[this.$store.taxYears.length - 2];
         }
-        txs = txs.filter(tx => {
-          return tx.assetEndingYears.findIndex(ey => ey == taxYear) > -1;
+        txs = txs.filter((tx) => {
+          return tx.assetEndingYears.findIndex((ey) => ey == taxYear) > -1;
         });
       }
       if (this.balanceGrouping == "Detailed") {
@@ -86,23 +87,23 @@ export default {
       }
 
       return Object.freeze(txs);
-    }
+    },
   },
   methods: {
     async load() {
-      const runningBalances = await getRunningBalances();
-      Vue.set(this, "runningBalances", Object.freeze(runningBalances));
-    }
+      const _runningBalances = await getRunningBalances(this.$store);
+      this.runningBalances = Object.freeze(_runningBalances);
+    },
   },
   async created() {
     await this.load();
-    store.onload = this.load();
+    this.$store.onload = this.load();
   },
-  destroyed() {
-    store.onload = null;
+  unmounted() {
+    this.$store.onload = null;
   },
   mounted() {
-    window.__vue_mounted = "PageTokenTransactions";
-  }
+    window.__vue_mounted = "PageRunningBalances";
+  },
 };
 </script>
