@@ -4,7 +4,6 @@
       :title="'Spending - ' + $store.taxYear"
       :data="filtered"
       :columns="columns"
-      @row-click="click"
     >
       <template v-slot:top-right>
         <filter-account-asset></filter-account-asset>
@@ -28,16 +27,14 @@
 </template>
 
 <script>
-import { store } from "../boot/store";
-import { actions } from "../boot/actions";
 import { getSpending, columns } from "../services/spending-provider";
-import Vue from "Vue";
+
 import TableTransactions from "src/components/TableTransactions.vue";
 import FilterAccountAsset from "src/components/FilterAccountAsset.vue";
 import {
   filterByAccounts,
   filterByAssets,
-  filterByYear
+  filterByYear,
 } from "../services/filter-service";
 
 export default {
@@ -49,8 +46,6 @@ export default {
       spendingGrouping: "Totals",
       spending: Object.freeze([]),
       columns: columns,
-      $store: store,
-      $actions: actions
     };
   },
   computed: {
@@ -64,7 +59,7 @@ export default {
       if (this.spendingGrouping == "Asset Totals") {
         for (const tx of txs) {
           let total = totals.find(
-            t => t.asset == tx.asset && t.taxCode == tx.taxCode
+            (t) => t.asset == tx.asset && t.taxCode == tx.taxCode
           );
           if (!total) {
             total = {
@@ -73,7 +68,7 @@ export default {
               amount: 0.0,
               fee: 0.0,
               gross: 0.0,
-              net: 0.0
+              net: 0.0,
             };
             totals.push(total);
           }
@@ -84,14 +79,14 @@ export default {
         }
       } else {
         for (const tx of txs) {
-          let total = totals.find(t => t.taxCode == tx.taxCode);
+          let total = totals.find((t) => t.taxCode == tx.taxCode);
           if (!total) {
             total = {
               taxCode: tx.taxCode,
               amount: 0.0,
               fee: 0.0,
               gross: 0.0,
-              net: 0.0
+              net: 0.0,
             };
             totals.push(total);
           }
@@ -103,26 +98,23 @@ export default {
       }
 
       return totals;
-    }
+    },
   },
   methods: {
-    click() {
-      //TODO add method editor popup
-    },
     async load() {
       const spending = await getSpending();
-      Vue.set(this, "spending", Object.freeze(spending));
-    }
+      this.spending = Object.freeze(spending);
+    },
   },
   async created() {
     await this.load();
-    store.onload = this.load();
+    this.$store.onload = this.load();
   },
-  destroyed() {
-    store.onload = null;
+  unmounted() {
+    this.$store.onload = null;
   },
   mounted() {
     window.__vue_mounted = "PageSpending";
-  }
+  },
 };
 </script>
