@@ -2,9 +2,8 @@
   <q-page class="" id="pageTokenTransactions">
     <table-transactions
       :title="'Token Transactions - ' + $store.taxYear"
-      :data="filtered"
+      :rows="filtered"
       :columns="columns"
-      @row-click="click"
     >
       <template v-slot:top-right>
         <filter-account-asset></filter-account-asset>
@@ -22,13 +21,12 @@
 import { store } from "../boot/store";
 import { actions } from "../boot/actions";
 import { getTokenTransactions, columns } from "../services/token-tx-provider";
-import Vue from "Vue";
 import TableTransactions from "src/components/TableTransactions.vue";
 import FilterAccountAsset from "src/components/FilterAccountAsset.vue";
 import {
   filterByAccounts,
   filterByAssets,
-  filterByYear
+  filterByYear,
 } from "../services/filter-service";
 
 export default {
@@ -40,7 +38,7 @@ export default {
       tokenTransactions: Object.freeze([]),
       columns: columns,
       $store: store,
-      $actions: actions
+      $actions: actions,
     };
   },
   computed: {
@@ -49,10 +47,10 @@ export default {
       txs = filterByAccounts(txs, this.$store.selectedAccounts, true);
       txs = filterByAssets(txs, this.$store.selectedAssets);
       if (this.onlyShowTracked) {
-        txs = txs.filter(tx => tx.tracked);
+        txs = txs.filter((tx) => tx.tracked);
       }
       return Object.freeze(txs);
-    }
+    },
   },
   methods: {
     click() {
@@ -61,18 +59,18 @@ export default {
     },
     async load() {
       const tokenTxs = await getTokenTransactions();
-      Vue.set(this, "tokenTransactions", Object.freeze(tokenTxs));
-    }
+      this.tokenTransactions = Object.freeze(tokenTxs);
+    },
   },
   async created() {
     await this.load();
     store.onload = this.load();
   },
-  destroyed() {
+  unmounted() {
     store.onload = null;
   },
   mounted() {
     window.__vue_mounted = "PageTokenTransactions";
-  }
+  },
 };
 </script>
