@@ -1,6 +1,12 @@
-
-module.exports = {
-  presets: [
-    '@quasar/babel-preset-app'
-  ]
-}
+module.exports = (api) => {
+  return {
+    presets: [
+      [
+        "@quasar/babel-preset-app",
+        api.caller((caller) => caller && caller.target === "node")
+          ? { targets: { node: "current" } }
+          : {},
+      ],
+    ],
+  };
+};
