@@ -11,8 +11,10 @@ const envparser = require("./config/envparser");
 const configuration = require("./package.json");
 const VERSION = configuration.version;
 const _env = require("dotenv").config().parsed;
+const ESLintPlugin = require("eslint-webpack-plugin");
+const { configure } = require("quasar/wrappers");
 
-module.exports = function(/* ctx */) {
+module.exports = configure(function (ctx) {
   return {
     // https://quasar.dev/quasar-cli/supporting-ts
     supportTS: false,
@@ -39,12 +41,13 @@ module.exports = function(/* ctx */) {
       // 'roboto-font-latin-ext', // this or either 'roboto-font', NEVER both!
 
       "roboto-font", // optional, you are not bound to it
-      "material-icons" // optional, you are not bound to it
+      "material-icons", // optional, you are not bound to it
     ],
 
     // Full list of options: https://quasar.dev/quasar-cli/quasar-conf-js#Property%3A-build
     build: {
       env: Object.assign({ appName: "CryptoTracker", VERSION }, _env),
+      devtool: "source-map",
       vueRouterMode: "hash", // available values: 'hash', 'history'
 
       // transpile: false,
@@ -62,21 +65,30 @@ module.exports = function(/* ctx */) {
 
       // Options below are automatically set depending on the env, set them if you want to override
       // extractCSS: false,
-
+      // extendWebpack(cfg) {
+      //   cfg.plugins.push(
+      //     new webpack.ProvidePlugin({
+      //       Buffer: ["buffer", "Buffer"],
+      //     })
+      //   );
+      // },
       // https://quasar.dev/quasar-cli/handling-webpack
-      extendWebpack(cfg) {
-        cfg.module.rules.push({
-          enforce: "pre",
-          test: /\.(js|vue)$/,
-          loader: "eslint-loader",
-          exclude: /node_modules/
-        });
-      }
+      chainWebpack(chain) {
+        chain
+          .plugin("eslint-webpack-plugin")
+          .use(ESLintPlugin, [{ extensions: ["js", "vue"] }]);
+        // const nodePolyfillWebpackPlugin = require("node-polyfill-webpack-plugin");
+        // chain.plugin("node-polyfill").use(nodePolyfillWebpackPlugin);
+        // const Buffer = require("buffer/").Buffer;
+        // chain.plugin("buffer").use(Buffer);
+      },
     },
 
     // Full list of options: https://quasar.dev/quasar-cli/quasar-conf-js#Property%3A-devServer
     devServer: {
-      https: false,
+      server: {
+        type: "http",
+      },
       port: 8080,
       open: false, // opens browser window automatically
       proxy: {
@@ -84,16 +96,16 @@ module.exports = function(/* ctx */) {
           target: "https://api.pro.coinbase.com",
           changeOrigin: true,
           pathRewrite: {
-            [`^/cbp-api`]: ""
-          }
-        }
-      }
+            [`^/cbp-api`]: "",
+          },
+        },
+      },
     },
 
     // https://quasar.dev/quasar-cli/quasar-conf-js#Property%3A-framework
     framework: {
       iconSet: "material-icons", // Quasar icon set
-      lang: "en-us", // Quasar language pack
+      lang: "en-US", // Quasar language pack
       config: {},
 
       // Possible values for "importStrategy":
@@ -109,7 +121,7 @@ module.exports = function(/* ctx */) {
       // directives: [],
 
       // Quasar plugins
-      plugins: ["LocalStorage", "Dialog", "Notify"]
+      plugins: ["LocalStorage", "Dialog", "Notify"],
     },
 
     // animations: 'all', // --- includes all animations
@@ -118,7 +130,7 @@ module.exports = function(/* ctx */) {
 
     // https://quasar.dev/quasar-cli/developing-ssr/configuring-ssr
     ssr: {
-      pwa: false
+      pwa: false,
     },
 
     // https://quasar.dev/quasar-cli/developing-pwa/configuring-pwa
@@ -137,30 +149,30 @@ module.exports = function(/* ctx */) {
           {
             src: "icons/icon-128x128.png",
             sizes: "128x128",
-            type: "image/png"
+            type: "image/png",
           },
           {
             src: "icons/icon-192x192.png",
             sizes: "192x192",
-            type: "image/png"
+            type: "image/png",
           },
           {
             src: "icons/icon-256x256.png",
             sizes: "256x256",
-            type: "image/png"
+            type: "image/png",
           },
           {
             src: "icons/icon-384x384.png",
             sizes: "384x384",
-            type: "image/png"
+            type: "image/png",
           },
           {
             src: "icons/icon-512x512.png",
             sizes: "512x512",
-            type: "image/png"
-          }
-        ]
-      }
+            type: "image/png",
+          },
+        ],
+      },
     },
 
     // Full list of options: https://quasar.dev/quasar-cli/developing-cordova-apps/configuring-cordova
@@ -170,7 +182,7 @@ module.exports = function(/* ctx */) {
 
     // Full list of options: https://quasar.dev/quasar-cli/developing-capacitor-apps/configuring-capacitor
     capacitor: {
-      hideSplashscreen: true
+      hideSplashscreen: true,
     },
 
     // Full list of options: https://quasar.dev/quasar-cli/developing-electron-apps/configuring-electron
@@ -191,16 +203,21 @@ module.exports = function(/* ctx */) {
       builder: {
         // https://www.electron.build/configuration/configuration
 
-        appId: "crypto-tracker"
+        appId: "crypto-tracker",
       },
 
-      // More info: https://quasar.dev/quasar-cli/developing-electron-apps/node-integration
-      nodeIntegration: true,
+      chainWebpackMain(chain) {
+        chain
+          .plugin("eslint-webpack-plugin")
+          .use(ESLintPlugin, [{ extensions: ["js"] }]);
+      },
 
-      extendWebpack(/* cfg */) {
-        // do something with Electron main process Webpack cfg
-        // chainWebpack also available besides this extendWebpack
-      }
-    }
+      // "chain" is a webpack-chain object https://github.com/neutrinojs/webpack-chain
+      chainWebpackPreload(chain) {
+        chain
+          .plugin("eslint-webpack-plugin")
+          .use(ESLintPlugin, [{ extensions: ["js"] }]);
+      },
+    },
   };
-};
+});
