@@ -5,20 +5,31 @@
   <q-form>
     <div class="row">
       <div class="col-12 col-sm-8">
-        <q-input v-model="address.name" label="Name"></q-input>
+        <q-input v-model="value.name" label="Name"></q-input>
       </div>
       <div class="col-12 col-sm-4">
-        <q-select v-model="address.type" :options="options" label="Type" />
+        <q-select v-model="value.type" :options="options" label="Type" />
       </div>
     </div>
-    <q-input v-model="address.address" label="Eth Acct"></q-input>
+    <q-input v-model="value.address" label="Eth Acct"></q-input>
   </q-form>
 </template>
 
 <script>
 export default {
   name: "FormAddress",
-  props: ["address"],
+  props: ["modelValue"],
+  computed: {
+    value: {
+      get() {
+        return this.modelValue;
+      },
+      set(value) {
+        this.$emit("update:modelValue", value);
+      },
+    },
+  },
+
   data() {
     return {
       options: [
@@ -30,10 +41,10 @@ export default {
         "Tax Deductible Donation",
         "Spending",
         "Expense",
-        "Income"
-      ]
+        "Income",
+      ],
     };
-  }
+  },
 };
 </script>
 
