@@ -6,8 +6,8 @@
     :data="data"
     :columns="columns"
     @row-click="click"
-    v-on="$listeners"
-    :pagination.sync="pagination"
+    v-bind="$attrs"
+    v-model:pagination="pagination"
     :rows-per-page-options="[0]"
     :style="{ height: tableHeight }"
     ref="transactionTable"
