@@ -44,9 +44,8 @@
       content-class="bg-grey-1"
     >
       <q-list>
-        <template v-for="(menuItem, index) in menuList">
+        <template v-for="(menuItem, index) in menuList" :key="index">
           <q-item
-            :key="index"
             clickable
             :to="menuItem.to"
             :active="menuItem.to === $route.path"
@@ -92,102 +91,101 @@ const menuList = [
     icon: "mdi-view-dashboard-variant",
     label: "Dashboard",
     separator: true,
-    to: "/"
+    to: "/",
   },
   {
     icon: "mdi-key-chain-variant",
     label: "Addresses",
     separator: false,
-    to: "/addresses"
+    to: "/addresses",
   },
   {
     icon: "mdi-door-open",
     label: "Opening Positions",
     separator: false,
-    to: "/opening-positions"
+    to: "/opening-positions",
   },
 
   {
     icon: "mdi-swap-horizontal-circle",
     label: "Chain Transactions",
     separator: false,
-    to: "/chain-transactions"
+    to: "/chain-transactions",
   },
   {
     icon: "mdi-swap-horizontal-variant",
     label: "Token Transactions",
     separator: false,
-    to: "/token-transactions"
+    to: "/token-transactions",
   },
   {
     icon: "mdi-file-swap",
     label: "Exchange Trades",
     separator: false,
-    to: "/exchange-transactions"
+    to: "/exchange-transactions",
   },
   {
     icon: "mdi-file-swap",
     label: "Offchain Transfers",
     separator: false,
-    to: "/offchain-transfers"
+    to: "/offchain-transfers",
   },
   {
     icon: "mdi-note-text-outline",
     label: "Running Balances",
     separator: false,
-    to: "/running-balances"
+    to: "/running-balances",
   },
   {
     icon: "mdi-cash-multiple",
     label: "Income",
     separator: false,
-    to: "/income"
+    to: "/income",
   },
   {
     icon: "mdi-credit-card-minus",
     label: "Spending",
     separator: false,
-    to: "/spending"
+    to: "/spending",
   },
   {
     icon: "mdi-calculator",
     label: "Capital Gains",
     separator: true,
-    to: "/capital-gains"
+    to: "/capital-gains",
   },
   {
     icon: "mdi-database-arrow-right",
     label: "Import",
     separator: false,
-    to: "/import"
+    to: "/import",
   },
   {
     icon: "mdi-database-arrow-down",
     label: "Tax Export",
     separator: true,
-    to: "/tax-export"
+    to: "/tax-export",
   },
   {
     icon: "mdi-cog",
     label: "Settings",
     separator: false,
-    to: "/settings"
+    to: "/settings",
   },
   {
     icon: "mdi-alert",
     label: "Backup",
     separator: true,
-    to: "/backup"
+    to: "/backup",
   },
   {
     icon: "help",
     iconColor: "primary",
     label: "Help",
     to: "/help",
-    separator: false
-  }
+    separator: false,
+  },
 ];
-import { store } from "../boot/store";
 import { actions } from "../boot/actions";
 import { processFiles } from "../services/file-handler";
 import { getRunningBalances } from "src/services/running-balances-provider";
@@ -201,9 +199,9 @@ export default {
       leftDrawerOpen: false,
       files: null,
       menuList,
-      taxYear: store.taxYear,
-      taxYears: store.taxYears,
-      $store: store
+      taxYear: this.$store.taxYear,
+      taxYears: this.$store.taxYears,
+      $store: this.$store,
     };
   },
   computed: {
@@ -213,26 +211,26 @@ export default {
         this.$store.addressesNeedsBackup ||
         this.$store.settingsNeedsBackup
       );
-    }
+    },
   },
 
   watch: {
-    taxYear: function(val) {
+    taxYear: function (val) {
       actions.setObservableData("taxYear", val);
     },
-    files: async function(val) {
+    files: async function (val) {
       if (val && val.length == 0) return;
       if (val && val.length) this.$store.importing = true;
       await processFiles(val);
       this.$store.importing = false;
       this.files = [];
-      if (store.onload) store.onload();
-    }
+      if (this.$store.onload) this.$store.onload();
+    },
   },
   async mounted() {
-    if (store.updated) {
-      await getRunningBalances();
-    }
-  }
+    // if (this.$store.updated) {
+    //   await getRunningBalances(this.$store);
+    // }
+  },
 };
 </script>
